// src/types/interfaces.ts
export interface Brand {
    id: number;
    name: string;
    slug: string;
    short_name: string;
    image: string | null;
    small_image: string | null;
    medium_image: string | null;
    large_image: string | null;
}

export interface Department {
    id: number;
    name: string;
    slug: string;
    image: string;
    description: string;
    status: string;
}

export interface ProductType {
    id: number;
    name: string;
    slug: string;
    short_name: string;
    image: string | null;
    small_image: string | null;
    medium_image: string | null;
    large_image: string | null;
}

export interface WebInfo {
    id: number;
    product_id: number;
    summary: string;
    description: string;
    is_splitted_with_colors: boolean;
    heading: string;
    meta_title: string;
    meta_keywords: string;
    meta_description: string;
    status: number;
}

export interface SizeDetail {
    id: number;
    size_scale_id: number;
    name: string;
    new_code: string;
    old_code: string | null;
    length: string | null;
    status: string;
}

export interface ProductSize {
    id: number;
    product_id: number;
    size_id: number;
    mrp: string;
    web_price: string;
    web_sale_price: string;
    detail: SizeDetail;
}

export interface ColorDetail {
    id: number;
    name: string;
    slug: string;
    short_name: string;
    code: string;
    ui_color_code: string;
    status: string;
}

export interface ProductColor {
    id: number;
    product_id: number;
    color_id: number;
    supplier_color_code: string;
    supplier_color_name: string;
    swatch_image_path: string | null;
    detail: ColorDetail;
}

export interface Filter {
    product_types: { id: string; name: string }[];
    sizes: { id: string; name: string }[];
    colors: { id: string; color_code: string }[];
    price: { min: number; max: number };
}

export interface PaginationProps {
    current_page: number;
    per_page: number;
    total: number;
    last_page: number;
}

export interface ProductImageProps {
    path: string;
    alt: string;
    product_color_id: number;
    is_default: boolean;
}

export interface ProductBase {
    id: number;
    slug: string;
    name: string;
    price: number;
    sale_price: number | null;
    default_image: string | null;
    brand: Brand;
    images: ProductImageProps[];
}

export interface Product extends ProductBase {
    article_code: string;
    manufacture_code: string;
    default_image: string | null;
    brand_id: number;
    department_id: number;
    product_type_id: number;
    price: number;
    sale_price: number | null;
    sale_start: string | null;
    sale_end: string | null;
    season: string | null;
    size_scale_id: number;
    min_size_id: number;
    max_size_id: number;
    department: Department;
    productType: ProductType;
    webInfo: WebInfo;
    specifications: any[];
    sizes: ProductSize[];
    colors: ProductColor[];
}